chore: dev tooling -- ESLint, CI, package validation, typos, flake 26.05
Stand up the quality tooling the project lacked:

- ESLint (flat config) with typescript-eslint and type-aware promise rules;
  fix what it found (floating/misused promises in `watch`, unnecessary
  assertions). The old `lint` script used a removed flag with no config.
- GitHub Actions CI: build, typecheck, lint, package-validate, unit tests,
  plus a typos spell-check job.
- `publint`/`attw` (`check:package`), which surfaced that the library
  shipped no `types`/`files` -- fix package.json `exports` accordingly.
- `typos` config + Nix dev-shell entry; bump the flake to nixos-26.05 and
  drop the removed `nodePackages` attribute.

+++ src/utils/content.ts
@@ -24,7 +24,7 @@ export function isContentEqual(
     return content1 === content2;
   } else {
     // Compare Uint8Array using native Buffer.equals() for better performance
-    const buf1 = content1 as Uint8Array;
+    const buf1 = content1;
     const buf2 = content2 as Uint8Array;
 
     if (buf1.length !== buf2.length) return false;
